Contact Form Tests: fix phpcs issues
Clean up the PHPCS issues in the test-class.grunion-contact-form.php file. There are so many
PHPCS issues in this file that the maximum number of errors is reached before the end of the file.
This cleanup will allow us to see PHPCS issues when adding new tests at the end of the file.